backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
python-dotenv==1.0.0
pydantic[email]==2.4.2
python-multipart==0.0.6